+++ README.md
@@ -1,3 +1,3 @@
 # graph_models
 
-Запуск `python3 ./<что-то из репозитория>` из корня.
+Запуск `python3 ./main.py` из корня.
