+++ downloader.py
@@ -21,14 +21,3 @@ def extract():
     os.system('rm ' + ARCHIVENAME)
     os.system('mv ' + TARFOLDER + ' ' + DATASETFOLDER)
 
-
-def get_samples():
-    with open(FOLDER + LABELS, 'r') as fp:
-        samples = json.load(fp)
-    images = {}
-    for image in samples['images']:
-        images[image['id']] = {'file_name': FOLDER + image['file_name'], 'annotations': []}
-    for ann in samples['annotations']:
-        images[ann['image_id']]['annotations'].append(ann)
-    return images
-
