+++ utils.py
@@ -1,4 +1,5 @@
 from lib_imports import *
+from consts import *
 
 
 def get_file_names(folder):
@@ -23,4 +24,15 @@ def create_path(path):
     for folder in folders:
         cur_path += folder + '/'
         create_folder_if_not_exists(cur_path)
+
+
+def get_samples():
+    with open(FOLDER + LABELS, 'r') as fp:
+        samples = json.load(fp)
+    images = {}
+    for image in samples['images']:
+        images[image['id']] = {'file_name': FOLDER + image['file_name'], 'annotations': []}
+    for ann in samples['annotations']:
+        images[ann['image_id']]['annotations'].append(ann)
+    return images
